move rerun logic to rerun.go
